presentation-api/controlling-ua/reconnectToPresentation_success-manual.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Reconnect to presentation success manual test</title>
<link rel="author" title="Marius Wessel" href="http://www.fokus.fraunhofer.de">
<link rel="author" title="Louay Bassbouss" href="http://www.fokus.fraunhofer.de">
<link rel="help" href="http://w3c.github.io/presentation-api/#dom-presentationrequest-reconnect">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="common.js"></script>

<p>Click the button below to start the manual test. Select a presentation device after the selection dialog is prompted. The test assumes that at least one presentation device is available. The test passes if a "PASS" result appears.</p>
<button id="startBtn">Start Test</button>

<script>
    // disable timeout for manual tests
    setup({explicit_timeout: true});
    async_test(function (t) {
        var startBtn = document.getElementById("startBtn");
        startBtn.onclick = t.step_func(function () {
            startBtn.disabled = true;
            var request = new PresentationRequest(presentationUrls);
            // start presentation
            request.start().then(t.step_func(function (oldConnection) {
                // presentationId will be used to reconnect to presentation
                var presentationId = oldConnection.id;
                var oldConnectionClosed = false;
                oldConnection.onclose = t.step_func(function () {
                    oldConnectionClosed = true;
                    // reconnect to presentation using presentationId
                    request.reconnect(presentationId).then(t.step_func(function (newConnection) {
                        assert_equals(newConnection.state, 'connecting',"connection should be in 'connecting' state");
                        assert_equals(newConnection.id, presentationId, "Ids of old and new connections must be equal");
                        t.done();
                    })).catch(t.step_func(function () {
                        assert_unreached("Reconnect to presentation promise rejected");
                    }));
                });
                // close connection. test continue after onclose event is fired.
                oldConnection.close();
                // the test fails if after 5s the close event is not fired
                t.step_timeout(function () {
                    !oldConnectionClosed && assert_unreached("Close event on old connection not fired");
                },5000);
            })).catch(t.step_func(function () {
                assert_unreached("Error on start presentation");
            }));
        });
    });
</script>
